+++ assets/test/short_answer/short_answer_authoring_test.ts
@@ -1,128 +1,44 @@
 import { ShortAnswerActions } from 'components/activities/short_answer/actions';
-import * as ContentModel from 'data/content/model';
-import { ShortAnswerModelSchema } from 'components/activities/short_answer/schema';
-import { ScoringStrategy } from 'components/activities/types';
-import produce from 'immer';
-
-const applyAction = (
-  model: ShortAnswerModelSchema,
-  action: any) => {
-
-  return produce(model, draftState => action(draftState));
-};
-
-function testFromText(text: string) {
-  return {
-    id: Math.random() + '',
-    content: {
-      model: [
-        ContentModel.create<ContentModel.Paragraph>({
-          type: 'p',
-          children: [{ text }],
-          id: Math.random() + '',
-        }),
-      ],
-      selection: null,
-    },
-  };
-}
-
-function testResponse(text: string, rule: string, score: number = 0) {
-  return {
-    id: Math.random() + '',
-    feedback: testFromText(text),
-    rule,
-    score,
-  };
-}
-
-export function testDefaultModel(): ShortAnswerModelSchema {
-
-  const responseA = testResponse('', 'input like {answer}', 1);
-  const responseB = testResponse('', 'input like {.*}', 0);
-
-  return {
-    stem: testFromText(''),
-    inputType: 'text',
-    authoring: {
-      parts: [
-        {
-          id: Math.random() + '',
-          scoringStrategy: ScoringStrategy.average,
-          responses: [responseA, responseB],
-          hints: [
-            testFromText(''),
-            testFromText(''),
-            testFromText(''),
-          ],
-        },
-      ],
-      transformations: [],
-      previewText: '',
-    },
-  };
-}
+import { defaultModel } from 'components/activities/short_answer/utils';
+import { ResponseActions } from 'components/activities/common/responses/responseActions';
+import { dispatch } from 'utils/test_utils';
 
 describe('short answer question', () => {
-  const model = testDefaultModel();
+  const model = defaultModel();
 
   it('has a stem', () => {
     expect(model).toHaveProperty('stem');
-  });
-
-  it('can edit stem', () => {
-    const newStemContent = testFromText('new content').content;
-    expect(applyAction(model,
-      ShortAnswerActions.editStem(newStemContent)).stem).toMatchObject({
-        content: newStemContent,
-      });
   });
 
   it('has input type', () => {
     expect(model).toHaveProperty('inputType');
   });
 
-  it('can edit feedback', () => {
-    const newFeedbackContent = testFromText('new content').content;
-    const firstFeedback = model.authoring.parts[0].responses[0];
-    expect(applyAction(model,
-      ShortAnswerActions.editFeedback(firstFeedback.id, newFeedbackContent))
-      .authoring.parts[0].responses[0].feedback)
-      .toHaveProperty('content', newFeedbackContent);
-  });
-
   it('has at least 3 hints', () => {
     expect(model.authoring.parts[0].hints.length).toBeGreaterThanOrEqual(3);
   });
 
-  it('can edit a rule', () => {
-    const newRule = 'input like {new answer}';
-    const response = model.authoring.parts[0].responses[0];
-    expect(applyAction(model,
-      ShortAnswerActions.editRule(response.id, newRule)).authoring.parts[0].responses[0])
-      .toHaveProperty('rule', newRule);
-  });
-
   it('can add and remove a response in text mode', () => {
-
-    const updated = applyAction(model, ShortAnswerActions.addResponse());
+    const updated = dispatch(model, ShortAnswerActions.addResponse());
     expect(updated.authoring.parts[0].responses[0].score).toBe(1);
     expect(updated.authoring.parts[0].responses[1].score).toBe(0);
     expect(updated.authoring.parts[0].responses[2].score).toBe(0);
-    expect(updated.authoring.parts[0].responses[0].rule).toBe('input like {answer}');
-    expect(updated.authoring.parts[0].responses[1].rule).toBe('input like {another answer}');
+    expect(updated.authoring.parts[0].responses[0].rule).toBe('input contains {answer}');
+    expect(updated.authoring.parts[0].responses[1].rule).toBe('input contains {another answer}');
     expect(updated.authoring.parts[0].responses[2].rule).toBe('input like {.*}');
 
-    expect(applyAction(updated,
-      ShortAnswerActions.removeReponse(
-        updated.authoring.parts[0].responses[1].id)).authoring.parts[0].responses).toHaveLength(2);
+    expect(
+      dispatch(
+        updated,
+        ResponseActions.removeResponse(updated.authoring.parts[0].responses[1].id),
+      ).authoring.parts[0].responses,
+    ).toHaveLength(2);
   });
 
   it('can add and remove a response in numeric mode', () => {
+    let updated = dispatch(model, ShortAnswerActions.setInputType('numeric'));
 
-    let updated = applyAction(model, ShortAnswerActions.setInputType('numeric'));
-
-    updated = applyAction(updated, ShortAnswerActions.addResponse());
+    updated = dispatch(updated, ShortAnswerActions.addResponse());
     expect(updated.authoring.parts[0].responses[0].score).toBe(1);
     expect(updated.authoring.parts[0].responses[1].score).toBe(0);
     expect(updated.authoring.parts[0].responses[2].score).toBe(0);
@@ -130,28 +46,11 @@ describe('short answer question', () => {
     expect(updated.authoring.parts[0].responses[1].rule).toBe('input = {1}');
     expect(updated.authoring.parts[0].responses[2].rule).toBe('input like {.*}');
 
-    expect(applyAction(updated,
-      ShortAnswerActions.removeReponse(
-        updated.authoring.parts[0].responses[1].id)).authoring.parts[0].responses).toHaveLength(2);
-  });
-
-  it('can add a cognitive hint before the end of the array', () => {
     expect(
-      applyAction(model, ShortAnswerActions.addHint()).authoring.parts[0].hints.length,
-    ).toBeGreaterThan(model.authoring.parts[0].hints.length);
-  });
-
-  it('can edit a hint', () => {
-    const newHintContent = testFromText('new content').content;
-    const firstHint = model.authoring.parts[0].hints[0];
-    expect(applyAction(model,
-      ShortAnswerActions.editHint(firstHint.id, newHintContent)).authoring.parts[0].hints[0])
-      .toHaveProperty('content', newHintContent);
-  });
-
-  it('can remove a hint', () => {
-    const firstHint = model.authoring.parts[0].hints[0];
-    expect(applyAction(model,
-      ShortAnswerActions.removeHint(firstHint.id)).authoring.parts[0].hints).toHaveLength(2);
+      dispatch(
+        updated,
+        ResponseActions.removeResponse(updated.authoring.parts[0].responses[1].id),
+      ).authoring.parts[0].responses,
+    ).toHaveLength(2);
   });
 });