+++ lib/oli/delivery/evaluation/parser.ex
@@ -32,6 +32,8 @@ defmodule Oli.Delivery.Evaluation.Parser do
   op_gt = ascii_char([?>]) |> optional(string(" ")) |> replace(:gt) |> label(">")
   op_eq = ascii_char([?=]) |> optional(string(" ")) |> replace(:eq) |> label("=")
   op_like = string("like") |> optional(string(" ")) |> replace(:like) |> label("like")
+  op_contains =
+    string("contains") |> optional(string(" ")) |> replace(:contains) |> label("contains")
 
   defcombinatorp(
     :string_until_rbrace,
@@ -61,8 +63,8 @@ defmodule Oli.Delivery.Evaluation.Parser do
 
   defcombinatorp(:component, choice([attempt_number_, input_, input_length_]))
 
-  # <operator> :== "<" | ">" | "=" | "like"
-  defcombinatorp(:operator, choice([op_lt, op_gt, op_eq, op_like]))
+  # <operator> :== "<" | ">" | "=" | "like" | "contains"
+  defcombinatorp(:operator, choice([op_lt, op_gt, op_eq, op_like, op_contains]))
 
   # <criterion> :== <component> <operator> <check>
   defcombinatorp(