+++ lib/oli/delivery/evaluation/rule.ex
@@ -48,6 +48,13 @@ defmodule Oli.Delivery.Evaluation.Rule do
     String.match?(eval(lhs, context), regex)
   end
 
+  defp eval({:contains, lhs, rhs}, context) do
+    String.contains?(
+      String.downcase(eval(lhs, context)),
+      String.downcase(rhs)
+    )
+  end
+
   defp eval(:attempt_number, context), do: context.activity_attempt_number |> Integer.to_string()
   defp eval(:input, context), do: context.input
   defp eval(:input_length, context), do: String.length(context.input) |> Integer.to_string()
@@ -68,7 +75,7 @@ defmodule Oli.Delivery.Evaluation.Rule do
 
   defp eval({:eq, lhs, rhs}, context) do
     left = eval(lhs, context)
-    right = eval(lhs, context)
+    right = eval(rhs, context)
 
     if is_float?(left) or is_float?(right) do
       {left, _} = Float.parse(left)