+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 ### Enhancements
 
 - Add ability to generate and download a course digest from existing course projects
-- Redesign check all that apply, multiple choice activities
+- Redesign check all that apply, multiple choice, short answer, and ordering activities
 - Merge activity editing into the page editor
 
 ## 0.11.1 (2021-6-16)
@@ -20,6 +20,7 @@
 - Fix an issue preventing deletion of projects whose names contain special characters
 - Fix an issue related to persisting sessions across server restarts
 - Fix an issue where modals and rearrange were broken in curriculum view
+- Fix an issue where toggling multiple choice answer correctness could cause submission failures
 
 ## 0.11.0 (2021-6-15)